spectrum scale coded in long (int capacity exceeded in some files)

+++ src/main/java/de/proteinms/omxparser/tools/OmssaViewer.java
@@ -529,7 +529,7 @@ public class OmssaViewer extends javax.swing.JFrame {
 
                     chargeString = chargeString.replaceFirst("\\+", "");
 
-                    int omssaAbundanceScale = tempSpectrum.MSSpectrum_iscale;
+                    long omssaAbundanceScale = tempSpectrum.MSSpectrum_iscale;
 
                     ArrayList<Double> currentRealMzValues = new ArrayList();
                     ArrayList<Double> currentRealAbundanceValues = new ArrayList();
